Beta.17 (#265)
* add new version

* add beta flag

* add edge extension

* update screenshots

* update edge id

* next page wip

* finish next page

* add comments

* update bookmarklet

* update error message

* extension/bookmarklet styling

* proper bookmarklet loading

* add data from bookmarklet

* update bookmarklet

* use intersection observer

* customize on all pages

* remove unused code

* update help page

* links to extensions

* fix bookmarklet

* update links

+++ extension/manifest.sh
@@ -18,6 +18,11 @@ then
   cp -r _source/ _public/opera/tmp
 fi
 
+if [ "$browser" == "edge" ]
+then
+  cp -r _source/ _public/edge/tmp
+fi
+
 
 applications=""
 
@@ -90,5 +95,14 @@ then
   echo "$manifest" > "_source/manifest.json"
   cd _public/opera/tmp
   zip -r ../opera_$version.zip ./*
+  rm -r ../tmp
+fi
+
+if [ "$browser" == "edge" ]
+then
+  echo "$manifest" > "_public/edge/tmp/manifest.json"
+  echo "$manifest" > "_source/manifest.json"
+  cd _public/edge/tmp
+  zip -r ../edge_$version.zip ./*
   rm -r ../tmp
 fi
